refactor: single out the model page components into a separate folder
